vendor/golang.org/x/net/html/testdata/webkit/tests23.dat

#data
<p><font tiles=4><font color=red><font tiles=4><font tiles=4><font tiles=4><font tiles=4><font tiles=4><font color=red><p>X
#errors
3: Start tag seen without seeing a doctype first. Expected “<!DOCTYPE html>”.
116: Unclosed elements.
117: End of file seen and there were open elements.
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <font>
|         tiles="4"
|         <font>
|           color="red"
|           <font>
|             tiles="4"
|             <font>
|               tiles="4"
|               <font>
|                 tiles="4"
|                 <font>
|                   tiles="4"
|                   <font>
|                     tiles="4"
|                     <font>
|                       color="red"
|     <p>
|       <font>
|         color="red"
|         <font>
|           tiles="4"
|           <font>
|             tiles="4"
|             <font>
|               tiles="4"
|               <font>
|                 color="red"
|                 "X"

#data
<p><font tiles=4><font tiles=4><font tiles=4><font tiles=4><p>X
#errors
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <font>
|         tiles="4"
|         <font>
|           tiles="4"
|           <font>
|             tiles="4"
|             <font>
|               tiles="4"
|     <p>
|       <font>
|         tiles="4"
|         <font>
|           tiles="4"
|           <font>
|             tiles="4"
|             "X"

#data
<p><font tiles=4><font tiles=4><font tiles=4><font tiles="5"><font tiles=4><p>X
#errors
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <font>
|         tiles="4"
|         <font>
|           tiles="4"
|           <font>
|             tiles="4"
|             <font>
|               tiles="5"
|               <font>
|                 tiles="4"
|     <p>
|       <font>
|         tiles="4"
|         <font>
|           tiles="4"
|           <font>
|             tiles="5"
|             <font>
|               tiles="4"
|               "X"

#data
<p><font tiles=4 id=a><font tiles=4 id=b><font tiles=4><font tiles=4><p>X
#errors
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <font>
|         id="a"
|         tiles="4"
|         <font>
|           id="b"
|           tiles="4"
|           <font>
|             tiles="4"
|             <font>
|               tiles="4"
|     <p>
|       <font>
|         id="a"
|         tiles="4"
|         <font>
|           id="b"
|           tiles="4"
|           <font>
|             tiles="4"
|             <font>
|               tiles="4"
|               "X"

#data
<p><b id=a><b id=a><b id=a><b><object><b id=a><b id=a>X</object><p>Y
#errors
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <b>
|         id="a"
|         <b>
|           id="a"
|           <b>
|             id="a"
|             <b>
|               <object>
|                 <b>
|                   id="a"
|                   <b>
|                     id="a"
|                     "X"
|     <p>
|       <b>
|         id="a"
|         <b>
|           id="a"
|           <b>
|             id="a"
|             <b>
|               "Y"
